driver phenom: rely on more trustable events to detect the sample loader as reached its target
When changing to SEM or NavCam positioni, instead of counting the number
of SEM images or NavCam images received, we wait for the operational
mode event. This way, it should be less dependent on the exact
behaviour of the Phenom firmware.

Also add timeouts, so that we don't block _forever_ when the hardware
doesn't behave as expected.

Also, this should help in case the NavCam fails to produce images (which
happens sometimes).