driver phenom: remove phenom_gui management
It was a trial, and it turned out to not be useful.
The main goal was to stop the Phenom GUI, which caused interference with Odemis.
Currently, the solution has been to fix the Phenom GUI so that it doesn't interfere in the first place.

Anyway, the released Phenom firmware doesn't even support this special way to
stop the GUI, so it's really useless to try.