acq stitching tiledacq: do not go back the initial position
It typically doesn't help anyone to move the stage to the first tile.
The stage could have been at an position before, and maybe the next
time we'll need the stage it will be near the last position.
So it just causes extra time, extra use of the stage, for no obvious
benefit.

If the caller really care about going back to the original position,
it should take care of saving and moving back itself.