driver orsay: add host info as roattribute
Temporarily, this information will have to be provided by the driver in
order to be able to access the server directly from the acq modules
(for milling and aperture control).
Eventually, we should provide an API directly in the orsay driver,
but anyway it still useful to be able to check what is the host setting.