added and changed code according to comments of reviewers.
changed move with gauge tracking in the GUI to a generic class in
/util/driver.py.
added support for future cancelling for moves with gauge tracking.

made a few changes to the sparc2-fslm-sim.odm.yaml file.

changed switching to a different mode in alignment tab to not move
to a position if requested axis is not in axes definition.