[feature] Extend hamamatsu streak camera driver to support C16910
The Hamamatsu C19610 is the new "universal" streak cam.
It's mostly the same as the previous streak cam, but has exchangeable
streak units (and delay boxes). As they have slightly different
settings, adjust the component VAs based on which device is connected.

Also add support to start without readout camera. Mainly to be able to
test using just the Hamamatsu simulator, which do not provide a camera.

Also allow to not have a delaybox, as in some configurations that's not
controlled by the Hamamatsu PC.

Also update regularly the VAs, so that if the user modifies the values
via the HPDTA, the changes are reflected in Odemis.

Also add an option to specify a INI file to start HPDTA with. This
indirectly allows to select a hardware profile.