[config] adjust categories of application launchers
Odemis was set to "Utilities", and the rest to nothing at all.
Instead, place Odemis and the viewer in Science (eg, with ImageJ),
and all the extra tools in system tools.

This is not used by default in Ubuntu dash, but if one uses the
"application" menu, these categories are used.