[feature] add support to backend for child component to rely on a power supplier
In case a component has children, it can be that each children needs a
power supplier independently. That's for instance the case for some of
the detectors connected to the DAQ board.
In such case, the backend need to first wait for the power supplier to
be instantiated before starting the creating of the "parent" component.