[feature] add support to backend for child component to have depencencies or power supplier
A component can only be initialised if its dependencies are ready.
If it has children, the dependencies of the children must also be ready.
That wasn't taken into account.

Note that no microscope file ever had such complexity... until
yesterday.

=> Also check for the dependencies of children component (including the
"power supplier" which is just a special type of dependency)