[doc] andorshrk: ky328 focus apparently can only go to 549 steps
The driver reports 550 steps, but tests on the actual hardware show that
the actual maximum position supported is 549. Otherwise, it raises an
error like:
Call to ShamrockSetFocusMirror failed with error 20201: COMMUNICATION ERROR

We could try to make it super clever... but for now we only support one
such device. Also, maybe future version of the SDK might be fixed, which
could make the code really complicated. Instead, the workaround is to
just limit the range using the "rng" parameter (typically, via the microscope
file).