[feature] keysight driver now check values
After changing a setting, check that it went fine, and if not, raise
an exception with the error reported by the device.

After changing a setting, read the accepted value, and use this "actual"
value as value for the VA.

When starting, check that the device looks like one supported.

Update the simulator to simulate these "read" commands.

Reguarly read the settings, in case they were changed by the user.