[fix] omicronxx handle device reporting 0xff bytes in the GSI command
Normally the GSI answer looks like (wl in nm) § int (power in mW).
However, for one source of one device, the answer is:
(wl in nm) 0xff 0xff § int (power in mW)

That's not defined in the documentation, but the official Omicron program
handles it fine... so we now also handle it fine.