[feat] Pause key also set detectors to "safe mode"
For now the [Pause] key (and "Stop all axes" menu) would only stop the
movement of the actuators, in order to bring the hardware to a safe
state. However, we can do more than that. Some detectors can be put a
safe state, to handle cases where too much "energy" is received. For
instance, the streak-cam has a shutter and gain which can be adjusted
for safety.

=> Introduce a new generic function to place "all detector" to a safe
mode... and for now add support for one: the streak-unit.

Also add a sort of callback to the tabs, to know when this function was
triggered. Will be used in a futher commit to handle better GUI updates
in such case.