[fix] SPARC hardware synchronized acquisition sometimes missing pixels
Sometimes the CCD takes quite a long time to get ready (ex, when the
readout rate is changed). Typically it takes 0.1, but in this cases, it
can take more than 1s. So the "large" margin we had put of 1s is not
enough. This meant the e-beam would start scanning before the CCD reacts
to the trigger, missing the first few pixels.

2s should be good enough for really every case?