[fix] semnidaq needs special case when only one image_ttl
The nidaqmx API apparently handles 1 DO channel differently from N
DO channels: with 1 channel, it expects the data to be boolean, while
for more than 1 channel, it should be a int32 (bitmask).

In practice, that meant that if the user configure the semnidaq driver
with just a single image_ttl (eg, "pixel" on a single port), then the
acquisition would fail.

=> Treat this case specifically, by requesting numpy a bool array
instead of uint32.