[fix] semnidaq tried to unregister callback although it was not registered
To keep the code simple, the driver would always unregister the event
callbacks at the end of an acquisition. However, with the latest version
of the NIDAQmx, this causes an error.
So now we have to be more careful, and only unregister the callback if
it was registered at the beggining of the acquisition.