[fix] andorcam2 don't call PrepareAcquisition, as it sometimes fails
The idea of PrepareAcquisition() was that it'd help the camera to start
more quickly when receiving a trigger. However, on the Newton without
software triggers, sometimes the camera would reject the call, because
it thinks it's not done with the acquisition yet (although it did
receive the frame already). As it's not clear whether it was ever
helpful, let's just drop this "potential improvement", and make sure
it's reliable.