[fix] semnidaq: improve NI task buffer handling
Clearly separate the HW buffer used by the NI-DAQmx kernel driver from
the "chunks" that the data is read/written by the Odemis driver. This
allows to ensure that the HW buffer is at least 2x bigger than the
chunks (when the data is continuous), which reduces the chances of
"data starvation" which happened in some cases.

Also fix an issue that caused the hardware to sometimes fail starting
the AI task. It seems there is bug (or undocumented feature) in the SDK
or hardware which makes the AI task very slow to start if a the DO task
has less than 4 samples. In practice, this could happen in spot mode.
=> just duplicate the data in such case (and as the task is repetitive,
it's completely equivalent).