[feature] picoquant: add support for PH330
The PicoHarp 330 is half-way between the PicoHarp 300 and the HydraHarp
400. => Reuse most of the code.

Also add support to change dynimically the trigger level and
zero-crossing level, for all devices. Still keep the old arguments for
the old devices allowed, to not break the current installations.

Also refactor the test cases to share as much code as possible.