[fix] odemisd: handle metadata for monochromator with both spectrograph and filter
With a monochromator, both the filter (in front of the spectrograph),
and the spectrograph center wavelength + output slit opening have effect
on which wavelength is observed. Until now, move on any of the axes
would update MD_OUT_WL... but only the latest move would define the
value.
That's especially annoying if the lastest move is to move the filter to
"pass-through" (very common). In this case, the MD_OUT_WL would be empty
(incorrectly).

The right way to compute MD_OUT_WL is to take into account all the axes
together. Essentially, taking the intersection between the filter
wavelength range and the output slit opening range.