[fix] hdf5: reduce log level for non useful metadata
If a metadata is considered not useful due to the state (eg ST_INVALID),
it would show a INFO warning.
A single HDF5 file can have hundreds of such metadata (because it the
standard way to indicate the value is unknown). That can cause a flood
of INFO log, absolutely useful.

=> Only log at the DEBUG level for these cases.