[fix] driver tescan: block a bit of time after unblanking.
Disabling the blanker takes a little bit of time. Typically, in Odemis,
this is done just before acquiring. However, the Tescan function immediately
returns, even if the beam is not ready yet.
So explicitly wait a bit, to make sure that the e-beam is ready when
Odemis needs to acquire an SEM image.

Note that we tested also the external mode, and it seems much faster,
not causing any image distortion if changing just before acquisition.