[feat] limit the maximum dwell time of the time-correlator
All detectors have their dwell time limited in the GUI (as some hardware can
support very long dwell/exposure times that make no sense for a user).
That is, except the time-correlator (for the SPARC labcube) which wasn't
limited. The picoquant hardware supports up to 4.5 days, which is really useless
in our use-case. => Limit to 1000 s, which is also the (arbitrary) limit of the
e-beam dwell time.