[feat] SEMTemporalMDStream now use synchronizing event on the time-correlator
The picoquant driver used to not support synchronization, so the
acquisition code had to start/stop explicitly the time-correlator
acquisition for each pixel. Now that the driver supports
synchronization, we can use the same pattern as the standard SPARC
acquisition.

This brings 2 advantages:
* avoids opening/closing "things" like the shutter and blanker for each
  pixel. This used to cause ~1s overhead per pixel.
* makes the code closer to the standard code, so we are one (little)
  step closer to merging this code into the standard code.

Also adjust the test cases to still work, now that the acquisition is
much faster.