[fix] METEOR: avoid aliasing in image of camera with large resolution
If the camera native resolution is larger than the viewport (> 1.9x), the image displayed
on the screen might have aliasing artifacts. (Eg, Andor Sona, with the METEOR v2).
To avoid this, we interpolate the content. The only drawback of interpolation is that it
lowers the frame rate, which can be annoying when manually focusing, so we don't activate
it by default.