[feat] acqmng: Make fold SPARC stream a little more lax
Dwell time is a float. In the GUI, it's rounded to 3 digits. Also, the
streams might adjust the value entered by the user to the value really
accepted by the hardware. Having precisely the same dwell time can be
really challenging. In addition, it's not a big deal if they are not all
the same, as at acquisition, one value will be picked (and that will
work).

=> Match dwell times if they are equal within 1%.