[fix] IndependentEBICStream now report the detector dwell time
The dwell time set by the user was used as-is to set the detector dwell
time, and then an extra time was added *if* playing to update the dwell
time. So sometimes the dwell time showed the detector dwell time and sometimes
the scanner dwell time (which is a little longer).

Change to always report the detector dwell time. Now that the ephemeron
driver .clip() return the acceptable value, changing the dwell time
works exactly the same when paused and playing.