[fix] XmlResourceHandler now explicitly handle the hidden flag
In the XRC files, each component may have a "hidden" flag. Up to
wxPython 4.0.x, which had wxWidgets 3.0, this flag was handled
automatically by .SetupWindow().

However, commit 7baf1ce2f (Don't show windows with "hidden" XRC attribute at all)
https://github.com/wxWidgets/wxWidgets/commit/7baf1ce2fd9bcf21e2bc336de46dd95fecf7a56b
changed this behaviour, and now every class has to do it (or not) in
DoCreateResource().

This makes sure it also happens on our own classes. This fixes issues
where GUI widgets would be incorrectly shown at init when using wxPython
4.1+ (for instance, on Ubunut 24.04).

Note that it always run, even on wxPython 4.0, but that's not an issue,
as it just makes the hidden flag a tiny bit more efficient.