[fix] SPARC2StreakCameraTestCase could sometimes fail
The simstreakcam simulator is not very good at simulating quick
stop/start. In some cases, it may not noticed it was stopped, and
immediately send an old image.

This confuses the test case. We could try to fix the simulator, but it's
simple to workaround it. The real driver (hamamatsurx) is not prone to
such issue.