[feat] GUI SPARC: at init warn if the interlock is triggered at start
During normal operation, if the laser interlock is triggered, a message
is shown, and the ebeam is blanked. Until now, nothing happened at init,
if the GUI starts while the interlock is already triggered.

With this change, in such corner case, the GUI will also show a
notification message, to remind the user something is "unusual".
However, the blanker is not activated, because it could be also cause
some confusion, or annoyance, if just the start of the Odemis GUI
blanks the beam while the user is using the SEM, and might have been
doing so safely for a while.