[fix] adjust timing for independent EBIC detector acquisition
Now use the same timing as in the live mode. In practice, this makes the acquisition
a lot more reliable.

Ideally, we would have a mechanism to read this timing from the Odemis
HwComponent, or have a synchronization mechanism. However, for now, we
only support one type of "independent detector", which is the Ephemeron
MightyEBIC. So let's not over-engineer the fix. This can be improved
once we have to support other types of detectors.