[ci] specify python version for linting
Setting up Python, it would warn like this:
Warning: Neither 'python-version' nor 'python-version-file' inputs were supplied. Attempting to find '.python-version' file.
Warning: .python-version doesn't exist.
Warning: The `python-version` input is not set.  The version of Python currently in `PATH` will be used.

Probably, that means it runs flake8 with whatever version of Python is
availble on the latest Ubuntu, but it's nicer to use an explicit
version, indeed.