[fix] andorcam3: limit the usage of Flush()
Flush() is used to discard all buffers already queued. In theory, calling Flush()
multiple times in a row should have no extra effect. So for "safety", we
used to call it before starting the acquisition.
However, on the Sona, with the current version of the Andor SDK3,
one call every ~500 in such condition causes the whole process to freeze.
If this happens, the only way to restore functionality is to restart
Odemis.

In practice, this caused time-outs in the acquisition of tiles or
z-stacks.

=> Never call Flush() when starting the acquisition, and always call
Flush() after stopping the acquisition.