[feat] andorcam3: improve Sona support
Add .frameRate VA to allow limiting the frame rate. This is helpful with
the Sona, as it can generate ~80 FPS at full res, but only transmit ~40
FPS on USB3. This means that by default, when an short exposure time is
used in live mode, the image stream get slowly delayed, until a timeout
happen. For such case, limiting the frame rate to 40 FPS allows to have
a smooth live stream.

To handle better the cases of timeout due to high frame rate, the
timeout duration is more finely computed too. That limits the freeze in
such case.

Also forbid changing the readout rate on the Sona, as it's directly
linked to the gain. That was confusing the users. Now, it's
automatically detected that each gain has its own readout rate, and it's
set as read-only VA.