[feat] acq: introduce functions to generate vector scan
New functions that allow to essentially scan the same way as the
standard e-beam scanner, but with the positions generated at the
"higher-level". They should result in exactly the same behaviour of the
e-beam as the standard scan, but with more flexibility.
The first (and only one) flexibility introduced is the support for
rotation (around the center of the RoI).