[feat] semnidaq: while readying the hardware for a task, let the e-beam move the first pixel
This "trick" is already used in the semcomedi driver. This gives extra
time for the e-beam to move to the first pixel. That can be useful for
when scanning very small regions: the fly back time is very tiny, but
for the very first frame, the e-beam still have to move from the park
position to the first pixel, which might require a lot longer.