[fix] bug_reporter: drop mix of wx.EXPAND and  wx.ALIGN_*
wx.EXPAND is supposed to force the sized widget inside a sizer to use
whole the space. So wx.ALIGN_* is never useful. The new version of
wxPython raise an error with wx.ALIGN_RIGHT and wx.ALIGN_BOTTOM. However
for wx.ALIGN_LEFT and wx.ALIGN_TOP, not because they actually do...
nothing (they are == 0), as they are the default value.

Still, having them there is confusing, so drop them.