WIP: semnidaq stop acquisition faster (attempt)
Currently, there is up to 100ms of delay between unsubscribing
and actually stopping the acquisition. That's an issue on the SPARC.
This attempts improving it by stopping the task (not working?!) and
waiting using an event before actually reading the data.

An alternative would be to be able to quickly switch to another task,
without stopping the previous one.