[feat] Tescan posture manager: make trajectory for posture switch safer
Until now there was quite some assumption on which Z is lowest depending
on the posture. As we add more posture, and have the METEOR2 which is
situated in a different place in the chamber, it's much harder to always
be certain the Z move is safe... in particular, moving from SEM to FM
could cause Z to go too high.

In order to make the posture switch safer, we now always first moves the
Z axis 5 mm below the lowest point, and from there, do the expected
moves, and finish by moving to the target Z.