[fix] Posture Manager sample stage should never move rotational axes
When converting from sample stage to stage bare coordinates, all the 5
axes are computed. So that 5 axes move was requested to the underlying
driver. However, on some hardware, such as on Tescan, asking to move an
axis, even if it's in the same position does cause a move (back and
forth). So, let's be more explicit, and only request to move the linear
axes, as for a given posture, we know the rotational axes never need to
move.