[fix] GUI: don't run workaround for wx bug on new wxPython
fix_static_text_clipping() is to workaround a but in wxPython v4.0, on
GTK. It's been fixed in wxPython v4.1. Although the workaround shouldn't
have any effet the the newest version, it does take extra time, and could
maybe in the future cause adverse effect. So it's better to just
explicitly do nothing with newest versions (eg, on Ubuntu 24.04).