[config] METEOR Tescan FIBSEM was missing some components
Optical lens component was missing... which was causing surprisingly
little issues in Odemis. Still, that meant the CCD image had no
MD_PIXEL_SIZE, which is required for some parts.

Add the SuperZ support, to test on a MIMAS too.

Adjust the filter positions to make them more realistic, and make the
switch faster.

The component names have changed, so the names in the "affects" should also be updated.