[feat] plugin photo_det_live: new plugin to add a "live" view of the photo-detector count
When adjusting the alignment of the fiber when a labcube is coupled
after the spectrograph, seeing the detector count over time can be useful.
Otherwise, the only view possible is the immediate count (via odemis
live). As this alignment is normally done only once per installation,
and there is only one installation, it's currently just a plugin. If it
needs to be done more frequently, we could add a new alignment mode to
the alignment tab.