[fix] GUI toolbar incorrect positionning when screen has High DPI scaling
When the scaling of the scaling is not 100%, the bitmap size is
automatically adjusted (ie, rescaled proportionally). However, the
toolbar used this information to adjust the toolbar size, which is in
"DIP" (Device Independent Pixel), meaning the value should be passed
assuming it's not scaled.
This caused the toolbar to be shown too large when the scaling > 100%.

Mostly noticeable on Ubuntu 24.04, with multiple screens,
where the scaling is automatically adjusted based on the screen DPI.