[fix] SPARC GUI: don't prevent acquisition when ebeam-gun-exciter interlock is open
As much as the light (aka laser) interlock can indicate a safety issue
(because it can mean the SPARC optical path is opened), an acquisition with
with the ebeam-gun-exciter interlock open would at worse mean that it
would acquire a dark measurement. More over, in many cases, the user
might not even want to be using the gun exciter. So no need to prevent
the acquisition in such case.