[fix] Tescan Posture Manager: FM sample stage move should depend on SEM rotation
The SEM scan rotation has an indirect effect on the FM movement: the FM
image is always transposed to match the SEM (and FIB) image, by
adjusting the microscope file. So if there is 180° scan rotation, the
FM relative moves need to be inverted.

This also ensure that old systems (where there is no scan rotation) do
still move in the right direction when using the new sample stage code.